minor refinement to similarity logic
When an origin entity contains a single-character word, require it to be at the start of a word in the destination entity, rather than anywhere. i.e. "hepatitis b vaccine" previously would be a child to "hepatitis c" because of the c in vaccine, but now no longer is.

Note that there's many other refinements that may work well for some environments but cause collateral damage elsewhere. This this felt generic enough to be useful in weeding out some false positives, but not too blunt in causing false negatives.